requirements.txt
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
fastapi==0.94.0
ftfy==6.1.1
grpcio==1.47.5
grpcio-tools==1.47.5
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
matplotlib-inline==0.1.6
mmh3==3.0.0
nest-asyncio==1.5.6
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.6
Pygments==2.14.0
pymilvus==2.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
setuptools==67.6.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.4.0
urllib3==1.26.15
uvicorn==0.21.0
wcwidth==0.2.6
wheel==0.38.4
zipp==3.15.0
